Improves naming coherence of method names, attributes, args & query symbols
The aim is to get rid of the ambiguous term "resource". My understanding of
the exist-db's docs is that conceptually a resource is a file of which all
(valid) XML files are considered documents. Only these are of interest here.
Also the notion of their XQuery function names that use the term "absolute"
lack a complementing "relative", so this term is bad design.

Also two remaining XQuery functions are handled with templates. With an attempt
to encapsule repeated expressions in string constants.

Some changes are emitting deprecation warnings, pointing to renamed equivalents.
Others just bang, they're easy to fix.

Docstring updates are deferred.

With these changes, also minor improvements such as "internal" attribute
protection are applied.